Created a separate FollowButton component and added it to both the candidate view and recommendations tab.
